% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barma.r
\name{fit.barma}
\alias{fit.barma}
\title{Model fit function}
\usage{
fit.barma(
  y,
  ar = NA,
  ma = NA,
  link = "aoz",
  diag = 1,
  h = 6,
  X = NA,
  X_hat = NA,
  resid = 1,
  lambda = 1
)
}
\arguments{
\item{y}{Time series}

\item{ar}{Autoregressive order}

\item{ma}{Moving-average order}

\item{link}{Link function}

\item{diag}{Output}

\item{h}{Lags for forecast}

\item{X}{Covariates}

\item{X_hat}{Observations of covariates used in forecast}

\item{resid}{Residuals type}
}
\description{
Funcao para ajuste do modelo BARMA com funcao de ligacao arbitraria (aranda, logit, probit, cloglog)
}
