% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.barma.r
\name{simu.barma}
\alias{simu.barma}
\title{Geração de ocorrências BARMA}
\usage{
simu.barma(
  n,
  phi = NA,
  theta = NA,
  beta = NA,
  X = NA,
  alpha = 0,
  prec = NA,
  freq = 12,
  link = "aoz",
  lambda = NA
)
}
\arguments{
\item{n}{Sample size}

\item{phi}{Autoregressive parameter}

\item{theta}{Moving-average parameter}

\item{beta}{Covariates}

\item{X}{Observations of covariates}

\item{alpha}{Intercept}

\item{prec}{Precision parameter}

\item{freq}{Number of observations per unit time. E.g. 12 for monthly data}

\item{link}{Link function}

\item{lambda}{Aranda-Ordaz parameter}
}
\description{
Funcao para gerar ocorrências provindas do modelo BARMA com função de ligação arbitrária (aranda, logit, probit, cloglog)
}
